requirements.txt
torch>=2.0
datasets
gensim
scikit-learn
matplotlib
seaborn
pyyaml
tqdm
numpy
pandas
scipy
